backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database clients
weaviate-client==4.9.3
motor==3.3.2
pymongo==4.6.1

# LLM and embeddings
google-generativeai==0.8.3
httpx==0.26.0
voyageai==0.2.3
tavily-python==0.3.3

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3

# Data processing
numpy==1.26.3
pandas==2.2.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring and logging
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
